chip/npcx: Use DIRTY bit to check if virtual wire status has changed
The NCPX eSPI Module has a set of registers called VWEVSM. These
contain information about the various virtual wires connected via eSPI
to the host. One of these is a 'DIRTY' which will be set when the host
recognizes a change in the status of one of the virtual wires that the
particular VWEVSM register is responsible for.

This patch changes the way SCI# and SMI# are signalled over these
virtual wires; instead of a fixed pulse width, the code will instead
poll the DIRTY bit to watch for the change, and exit polling after the
a pulse width's worth of time. Because the code now verifies that the
host has recognized the change in signal level, the extension of the
deasserted state of SCI# and SMI# is removed, as we already should
know that the host has captured both the asserted and then deasserted
states of the respective VWs.

BUG=b:230366077
BRANCH=brya
TEST=attempt SCI-loss reproduction scenario from
b:227367177, and verify that no SCIs were noticed as missing.

Signed-off-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Change-Id: I87ae66c7af067bd14633edf04f6ee9dc5ffcde99