zephyr/usba: unconstify usb_port_enable if needed
The changes in commit 7dc9537d7a56dab267e4facb3c1398e1d0e44175 added
support in Zephyr for CONFIG_PLATFORM_EC_USB_PORT_ENABLE_DYNAMIC with
the same semantics as the equivalent option in the legacy EC, but failed
to change the definition in the shim code to remove the const qualifier.
This would cause code that believed it could write to the
usb_port_enable array to instead trigger undefined behavior.

BUG=b:231512828
TEST=Nereid no longer hangs at boot when CBI fw_config is set to 0
     (unknown sub-board)
BRANCH=none

Change-Id: I747c9f1c1f02847cc367da38da1d4c7a12037516
Signed-off-by: Peter Marheine <pmarheine@chromium.org>