chargesplash: Add host display command
Create a new "chargesplash" tool designed to show the display using
frecon.

The display is intentionally nothing but a text-based UI at the
moment.  Future changes will add pretty graphics and such, but the
core logic is here with this commit.

Right now, the display loop just uses polling at 250 ms to read the
charging state and power button state.  For a better user experience,
we should grab the power button input device and terminate early in
the future.  This could even enable us to slow the polling further, as
charging state is not something that changes at high frequency anyway.

BUG=b:228370390
BRANCH=none
TEST=with whole CL stack, use ghost image and activate chargesplash on
     brya

Change-Id: I238ac4ebc6c4ba7cc0d65638ebbb7143e417a8cc
Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>