TCPMV2: Send Alert Message on Power State Change
Some docks will have lights which are meant to reflect the current power
state of a Chromebook. This CL is to send an Alert message to partners
when power state changes. If the power state is important to the
partner, they can then request more information with a Get_Status
command.

BUG=b:226228461
TEST=Changed power state of device and observed Alert message in EC log.
BRANCH=None

Signed-off-by: Jameson Thies <jthies@google.com>
Change-Id: If46c0e3b40d0a17c9b746c4efa480696a6293e9c