PCHG: Don't wake up device if lid is closed
Currently, a PCHG device wakes up the system even with the lid is
closed. Since the lid is closed, the system would re-suspend after 30
seconds but this is a little wasting power.

This patch makes the PCHG subsystem prevent device an attach or detach
event from resuming the system if the lid is closed.

BUG=b:192098599
BRANCH=None
TEST=Suspend CoachZ by closing the lid. Verify device doesn't resume by
attacing or removing a stylus.

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>
Change-Id: I11b6fbf30270174048de215a27905524a04a40bf