ap_pwrseq: forbid setting state outside the thread
Setting the state machine state or any individual power signal from outside
the power sequencing thread can cause incorrect behavior because changes to
any signals can cause others to change and in turn wake the power
sequencing thread which will begin to manipulate the same state in
unpredictable ways.

This error manifests as hangs in the shell thread when running `apshutdown`
for instance, because shutdown hooks will be run concurrently from the
shell and power sequencing threads as the shell thread manipulates power
signals to effect a forceful shutdown. This change does not correct that,
but does make it an error that causes a visible crash when assertions are
enabled.

State machine initialization is reorganized somewhat to be somewhat
clearer and avoid exposing potentially unsafe functions (inlining them in
the place they are known to be safe to use), and the G3 exit flag is
converted to an atomic value because it may be manipulated from other
threads (and it is safe to do so).

BUG=b:229815313,b:231281133
TEST=apshutdown from RO console now crashes on assertion failure instead
     of hanging. Normal operation appears unaffected.
BRANCH=none

Signed-off-by: Peter Marheine <pmarheine@chromium.org>
Change-Id: Id57cd14d1277a6da33d3104801337e6567e4e1af