power: Add Support for PMAX and PBSS
For DTT Battery participant, PMAX and PBSS indicates
platform maximum power and the power battery steady
state or the maximum sustained power of the battery.
These values must be read from Fuel guage registers
and the ec memmap regions must be updated. Also, if
there is any change of 250mW in PMAX or 100mW in
PBSS, CB is notified with an sci event to indicate
the change. Also, as part of DBPT initialization,
it is required to initialize the system resistance
and minimum system voltage ideal for the board.

BUG=b:205928013
BRANCH=None
TEST=make buildall

Signed-off-by: poornima tom <poornima.tom@intel.com>
Change-Id: I1ddcefc64f150d3c31e82c57b1ff110809c1cd59