guybrush: Enable i2c PEC for battery fuel gauge
The fuel gauge could NACK during a read operation and then return
invalid data. The only way to detect this corruption is to enable PEC.
The config enables PEC globally, but the smart battery is the only
driver that sets the I2C_FLAG_PEC flag, so this change will only affect
the smart battery.

BUG=b:230520109
BRANCH=guybrush
TEST=Boot nipperkin, no errors observed, battery status is correct

Change-Id: I50e60350ccff7e58f170c24921bc442eea0af817
Signed-off-by: Rob Barnes <robbarnes@google.com>