bc12: split task handler in init and event
Split the bc12 internal API from a task loop handler to an init and
event handler pair of functions, moving the loop on the caller.

This is to allow handling the events from a single task on a followup
patch.

Current driver behavior should be unchanged, just bringing the loop and
wait in the generic code. For a couple of drivers (bd9995x and rt946x)
this was not quite possible as they would need some more work to fit in
that model, but then those two runs in single task mode already anyway
(HAS_TASK_USB_CHG), so no point converting them to fit the model anyway,
just running the whole driver in the init() keeps the existin behavior,
code will never reach the outer loop.

BRANCH=none
BUG=b:226411332
TEST=make buildall
TEST=zmake testall
TEST=cq dry run

Signed-off-by: Fabio Baltieri <fabiobaltieri@google.com>
Change-Id: I8e28ad8b5b52b002587aefb68116a14b709729fe