chip/npcx: Poll VW DIRTY bit around SCI# and SMI# pulses
This patch changes the way SCI# and SMI# are signalled over their
virtual wires; before and after generating the low pulse, the code
will ensure that the VWEVSM DIRTY bit is not set, which should
correspond to the host acknowledging the change in VWIRE state.
Because the code now verifies that the host has recognized the change
in signal level, the extension of the deasserted state of SCI# and
SMI# is removed, as we already should know that the host has captured
both the asserted and then deasserted states of the respective VWs.

BUG=b:230366077
BRANCH=none
TEST=attempt SCI-loss reproduction scenario from
b:227367177, and verify that no SCIs were noticed as missing.

Signed-off-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Change-Id: I8cb72a92064ad98f2c5c49caa3c87fdfe7afa4d5