nissa: move nivviks private CBI fields to private DT overlay
The mainboard orientation field in CBI is only used by Nivviks, and the
system configuration is updated in https://crrev.com/i/4750272 to avoid
duplicating definitions for fields that are available to all boards. Move
the base inversion field to a private overlay since it now clearly applies
only to Nivviks.

BUG=none
TEST=builds
BRANCH=none

Signed-off-by: Peter Marheine <pmarheine@chromium.org>
Change-Id: Ibad19a14b91b597671468c317587a10e7cd588fe