power: Add support for PMAX in DBPT
For DTT Battery participant, PMAX indicates the
platform maximum power of the battery. This value
must be read from Fuel guage register and the ec
memmap region must be updated. Also, if there is
any change of 250mW in PMAX, CB must be notified
with an sci event to notify the change. The memmap
updation and check for the level change happens
every 10seconds. This CL takes care of this changes.

BUG=b:205928013
BRANCH=None
TEST=make buildall

Signed-off-by: poornima tom <poornima.tom@intel.com>
Change-Id: I7719617ae602ae87b14acb3d878f4c726769f240