zephyr: fix ioex_init function
The ioex_init function has to reconfigure the ioexpander chip. For the
Zephyr ioexpander driver, which is treated as a normal GPIO driver,
means we have to configure every single pin from the ioexpander chip.
Use the gpio_reset_port for that.

This change allows reconfiguring ioexpnader pins e.g. after the
ioexpander chip reset which is done during TCPC init for Brya board.

BUG=b:231086773, b:208181774
TEST=zmake testall && Make sure PD works on ports 0 and 2 on Brya
BRANCH=main

Signed-off-by: Dawid Niedzwiecki <dn@semihalf.com>
Change-Id: I8c446a4b35b03b9ed9574c6572a0daa80a9a7dee