OCP: Update OCP handling to match PD 3.1 spec
The PD 3.1 spec has added a new criteria that we must start
ErrorRecovery after three OCP events on a port.

Additionally, this change moves more of the OCP logic into the OCP
module itself and makes it more closely match the existing spec
instructions.  We now hard reset right away and queue up an alert
message for later when a new contract is in place.

BRANCH=guybrush
BUG=b:228403079,b:233364677
TEST=on nipperkin, OCP port repeatedly and observe expected resets and
Alert messages, and observe no Vbus is turned on after fourth OCP

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: Ibeccc942b1ab638ac93e6e74f0692202f51cf5bb