zephyr: Remove ec_version.h dependency on common/system.c
Per Jack, the only source in the Zephyr EC that should be including
ec_version.h is common/version.c. However, system.c also includes it to
use the CROS_FWID_MISSING_STR macro.

This CL removes the dependency and adds a function to version.c to allow
accessing this string.

BRANCH=None
BUG=None
TEST=zmake test test-drivers

Signed-off-by: Tristan Honscheid <honscheid@google.com>
Change-Id: I8bb8406823e7860162157c1baaa21440608f3eea