power: Add support for RBHF in DBPT
For DTT Battery participant, RBHF indicates the high
frequency impedance. This value must be read from
Fuel guage register and the ec memmap region must be
updated every 10 sec. Also, if there is any change
of 5mohm in RBHF, CB must be notified with an sci
event to notify the change. This CL takes care of
this changes.

BUG=b:205928013
BRANCH=None
TEST=make buildall

Signed-off-by: poornima tom <poornima.tom@intel.com>
Change-Id: Ib5e02f4830f0ebafb53dfe5050059f2006a93c93