power: Add support for VBNL in DBPT
For DTT Battery participant, VBNL indicates the
instantaneous no load voltage. This value must be
read from Fuel guage register and the ec memmap
region must be updated every 1 sec. Also, if
there is any change of 50mV in VBNL, CB must be
notified with an sci event to notify the change.
This CL takes care of this changes.

BUG=b:205928013
BRANCH=None
TEST=make buildall

Signed-off-by: poornima tom <poornima.tom@intel.com>
Change-Id: I32a4653050df5143b83d3a84c8baa59660745b7f