power: Add support for CMPP in DBPT
For DTT Battery participant, CMPP indicates the
discharge current capability of the battery.
This value must be read from Fuel guage register
and the ec memmap region must be updated every
10sec. Also, if there is any change of 100mA in
CMPP, CB must be notified with an sci event to
notify the change. This CL takes care of this
changes.

BUG=b:205928013
BRANCH=None
TEST=make buildall

Signed-off-by: poornima tom <poornima.tom@intel.com>
Change-Id: I21ffb96bafaef1ca82f4e80ff6710b0cc3c90380