corsola: set HPD level from the first PD Status
Respect the HPD level status from the first PD status.
This is to fix the case that the DP adapter is plugged to an
external monitor before plugged into the DUT, the HPD level was high
already, and we should turn the DP on and set active_aux_port to it.

BUG=b:232491934
TEST=Can output to external monitor when plugged the adapter with
     external monitor attached.
BRANCH=none

Change-Id: I98a8ed71e97d746b79de7e8995f5e32112d08387
Signed-off-by: Eric Yilun Lin <yllin@chromium.org>