Skyrim: Reset TCPC IO expander after chip reset
The TCPC may be reset to get it out of dead battery mode, at which point
we need to re-enable the IO expander ports.  Note we're still waiting on
an API in order to read and restore the values each port had from before
the chip reset.

BRANCH=None
BUG=b:225189538
TEST=on skyrim, boot from hibernate and ensure A0 works as long as the
reset was performed before entering S0

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: I61f07e5ad0d4e6bf1ff6ac65737394bcdb046b63