TCPMv2: Use ErrorRecovery rather than hard reset during FRS
While in the middle of an FRS, we should not allow any hard resets.  If
a hard reset is performed, we may mistake our own FRS Vbus for the
partner and end up stuck in Attached.SNK.

BRANCH=guybrush
BUG=b:230494120
TEST=on nipperkin, observe that ErrorRecovery is performed after RCP
event on FRS dongle unplug.  Observe that power loss to the FRS dongle
still results in a successful FRS

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: I081b27ec9552b15c2061fa97b57e8e6eb1552236