Enable sensor initialization to retry on failures
Separate the FP sensor init request from the execution to enable the
initialization logic to be run (or rerun) from fp_task. This enables
continually retrying init if a user's finger is on the sensor at startup.

A new console command has been added to enable reinitialization of the
FP sensor for development and testing purposes.

Mock and bep sensor logic were both updated to were also updated to
support the updated API and ensure tests keep running. Some device
tests may now fail due to changes to the initialization logic. Need to
investigate.

This commit includes some logic to avoid sleeping while the debugger is
connected. This should be removed and added in a different CL.

BUG=b:159748988
TEST=./test/run_device_tests.py -b dartmonkey
BRANCH=none

Signed-off-by: Bobby Casey <bobbycasey@google.com>
Change-Id: I4272c9f79f7be5457ba7e31be5670bd62bfb5284