zephyr:test: Add dock with power button test case
This CL adds a dock with power button test case and updates the source
emulator to support all the messages required for this currently enabled
in TCPMV2. The new tests will notify a series of power state changes to
the ec with hook_notify, and TCMPV2 will then have a
Alert->Get_Status->Status transaction with the emulated partner. This
test also includes a Get_Revision->Revision transaction with the partner
to enable the 7-byte status message defined in USB PD Revision 3.1.

BUG=b:232536412,b:233455723
BRANCH=NONE
TEST=zmake test --coverage test-drivers --clobber

Signed-off-by: Jameson Thies <jthies@google.com>

Change-Id: Ide0d60e9a2f7e7098af545a70ab5c417acb5d934