ec console:  Move from tostring to tobytes.
To string has been an alias for tobytes since python 3.2, in later
version of python ( 3.9 and beyond ) this alias is removed. https://bugs.python.org/issue38916

This is a save change as all of chrome dev is on 3.6 or better.

The error goes away but when using usb_console inside of a container
I get device busy errors as the servod is running.

BUG=b:234176994
TEST=built a servod with python 3.10 and tested usb_console.

Change-Id: I1b00b0dd96413490936e20aa28067324c9d8e9a5