sm5803: verify CHG_MON_REG settings when sinking
Incorrect settings for this register may be implicated in hardware
damage that has been observed, so add a runtime check for correct
configuration before enabling sinking. This should help debugging of
issues that may be observed and prevent damage.

CMD_CHARGER_DUMP is enabled on Nereid to assist in this investigation.

BUG=b:230712704
TEST=zmake build nereid, `charger dump` on earlier nereid shows charger
     register 5C reading expected 7A.
BRANCH=none

Change-Id: If81ec544b7b26ec20316cb7bc92eaa1770b1da77
Signed-off-by: Peter Marheine <pmarheine@chromium.org>