kinox: barrel jack priority first
According to the customer request, let the barrel jack supply all the
time and priority first and let the system can switch the power path on
S0.

Set the EN_PPVAR_BJ_ADP_L to high when power on the system by Type-C
adapter.

Disable the PPC sink mode and set the EN_PPVAR_BJ_ADP_L to low when
barrel jack plugin even system is in S0.

BUG=b:225769067
BRANCH=none
TEST=Power on the system by the Type-C adapter then power supply switch
to the barrel jack when plugin the barrel jack.

Signed-off-by: Matt Wang <matt_wang@compal.corp-partner.google.com>
Change-Id: If36f6fa86961b6b83885927d0ef616a6881985cb