power/mt8186: stay at S5 before PMIC is fully off
If we command the device to enter hibernate mode before
PMIC is fully off, then there will be power leakage from
AP_EC_WDTRST_L which is pulled up by PMIC.

On MT8186, the S5 was a transition state, it passed through
S5 and entered G3 before the PMIC is fully off.  In normal cases,
this won't be an issue that it takes long enough time to hibernate from
G3.  However, firmware_ECWakeFromULP force EC enter hibernation once
the DUT gets into the G3, and this hits the PMIC issue.

To fix this, we force the DUT staying at S5 until the EC_PMIC_EN_ODL
released (8 seconds), and then enter the G3 state.

BUG=b:233988259
TEST=stay at S5 until EC_PMIC_EN_ODL deasserted and then into G3.
TEST=Test following items:
 * Cold reset:
   $ dut-control cold_reset:on sleep:0.2 cold_reset:off
   Result: G3 -> S0
 * Long power press to shutdown:
   $ dut-control 'ec_uart_cmd:powerbtn 8200'
   Result: S0 -> S5 -> G3
 * Long power press to power-on but then shutdown:
   $ dut-control 'ec_uart_cmd:powerbtn 8200'
   Result: G3 -> S0 -> S5 -> G3
 * Short power press to power-on:
   $ dut-control 'ec_uart_cmd:powerbtn 200'
   Result: G3 -> S0
 * Console command: apreset
   Result: S0 -> S5 -> S0, AP reboots
 * Console command: apshutdown
   Result: S0 -> S5 -> G3
 * Lid open to power-on:
   $ dut-control lid_open:no sleep:0.2 lid_open:yes
   Result: G3 -> S0
BRANCH=none

Change-Id: Id5809ff5dc72e35a75d9b20b2013522610ae4eaf
Signed-off-by: Eric Yilun Lin <yllin@chromium.org>