zephyr: Remove cros_fwid from version output
Currently the ec version output prints the value of cros_fwid. However,
this string isn't set up for zephyr and is in the process of being
deprecated. This commit disables the Zephyr Kconfig for this feature by
default. This is done instead of removing it to pass a CQ test that
will fail because it identifies `CONFIG_CROS_FWID_VERSION` as an orphan
config.

Example EC console output:
```
22-05-26 13:07:39.852 Chip:     Nuvoton npcx9m3f 00160207
22-05-26 13:07:39.855 Board:    2
22-05-26 13:07:39.855 RO:       skyrim_v3.1.49995-ec:15e362,os:
22-05-26 13:07:39.858 RW:       skyrim_v3.1.49995-ec:15e362,os:
22-05-26 13:07:39.860 Build:    skyrim_v3.1.49995-ec:15e362,os:<...>
22-05-26 13:07:39.868           2022-05-26 13:04:44 <...>
```

BRANCH=none
BUG=b:232468323
TEST=Ran version command

Signed-off-by: Robert Zieba <robertzieba@google.com>
Change-Id: Ic80efaec46f67b971ccdde0441538545d3321d20