jinlon: moving buttons and switches to use MKBP
Switching config options to route buttons and switches to using MKBP
instead of 8042 driver.  This fixes the issue that we were seeing
where when the "Treat top-row keys as function keys" config switch was
enabled on Jinlon, the volume buttons (on the side of the device, not
the keyboard) were behaving like F11/F12 keys.

BUG=b:232107647
BRANCH=hatch
TEST=enable the "Tread top-row keys as function keys" switch
     make sure physical volume keys on the side of Jinlon device still
     work (we see the volume slider appear)

Signed-off-by: Shelley Chen <shchen@google.com>
Change-Id: I9a6aadc1921ac4a612cc3e7612955afdcc0f03c5