zephyr test: Improve SYV682x test naming
The UUT for this test is the SYV682x device driver, which works for part
versions A, B, and C. The (emulated) part in the devicetree is an
SYV682C, but the majority of the driver interface is unaffected by the
specific part. Update file, suite, and variable names to reflect this.

BUG=b:234886404
TEST=zmake testall
BRANCH=none

Signed-off-by: Abe Levkoy <alevkoy@chromium.org>
Change-Id: I8024239d4f4b4aa0793ea9b1e5a4b9df126a4009