zephyr: shim: ensure static shell priority matches the task one
Set the static SHELL_THREAD_PRIORITY config default and add a build
check to ensure that the atual value matches the one configured in the
list of intended task priorities.

This ensures that the shell task is running at the intended priority
even on projects that do not call shell_init_from_work().

BRANCH=none
BUG=none
TEST=zmake build brya

Signed-off-by: Fabio Baltieri <fabiobaltieri@google.com>
Change-Id: Ie11f18b1135aee04386105a7a87b5c4b1ef0cf8a