chip/stm32: Support SPI chip select across transactions
Under some circumstances, we want to assert SPI chip select, do a
transactions, and then hold the chip select while doing further
transactions (possibly depending on the result of the first
transaction.)  In particular, this happens on HyperDebug, which is a
uServo-like device, that receives instructions via USB on how to perform
SPI transactions.

This change adds code such that spi_transaction() and related functions
will make a note of the CS-level before performing the transaction, and
release CS after the transaction only if it was not asserted prior to
the transaction.

This adds negligible overhead to the common spi methods, yet I have
chosen for now to guard the added code by CONFIG_USB_SPI, such that only
servo-like devices (and the few EC chips that forward AP console on
GSC-less devices) would feel any added burden.

BUG=b:192262089
TEST=Use HyperDebug to interact with AnderiBoard via SPI
BRANCH=none

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: I55ffd2452f399735f8125f347255d33efc492a05