chip: Remove -march, since -mcpu itself is sufficient
According to the GCC documentation:

-mcpu specifies the name of the target ARM processor. GCC uses this name
to derive the name of the target ARM architecture (as if specified by
-march) and the ARM processor type for which to tune for performance (as
if specified by -mtune).

https://gcc.gnu.org/onlinedocs/gcc/ARM-Options.html

This means that unless we're overriding some setting with -march or
-mtune, there's no need to specify them. In our case, we aren't
overriding the default for -mtune that -mcpu sets, so make things
simpler by only specifying -mcpu.

Additional details in
https://community.arm.com/arm-community-blogs/b/tools-software-ides-blog/posts/compiler-flags-across-architectures-march-mtune-and-mcpu

BRANCH=none
BUG=b:145677491
TEST=./util/compare_build.sh -b all -j 120
     => MATCH

Signed-off-by: Tom Hughes <tomhughes@chromium.org>
Change-Id: I1999ef25642ed9e32b2e8eeb76d4ba7a23dc2b8a