ps8743: Fix mode to apply power optimization and value to write
Apply the power optimization if USB mode is on and DP mode is off. The
check of these modes are achieved by checking the Misc register.
Correct value to write to the mode register. As the optimization only
applies when only USB2 devices are connected, restructure the logic and
skip any operation if any USB3 device is connected.

BUG=b:229810117
TEST=Check the operation mode register (0x00) entering/leaving S3:
     (1) USB 3 device is connected, register value not changed with
         different type c side.
     (2) USB 2 device is connected, only the USB mode bit is altered in
         the register.
     (3) type c display is connected, register value not changed.
BRANCH=None

Change-Id: Ic85ee805ffea26bfa3fc5004c3d319cefcf2e68e
Signed-off-by: lschyi <lschyi@google.com>