ps8743: Fix mode to apply power optimization and value to write
Apply the power optimization if USB mode is on and DP mode is off. The
check of these modes are achieved by checking the Misc register.
Correct value to write to the mode register. As the optimization only
applies when only USB2 devices are connected, and should not turn on USB
mode if device is removed during S3, create enum and record status with
internal variable, then restructure the logic for suspend and resume.

BUG=b:229810117
TEST=Check the operation mode register (0x00) entering/leaving S3:
     (1) USB 3 device is connected, register value not changed with
         different type c side.
     (2) USB 2 device is connected, only the USB mode bit is unset at
         S3, and set at S0.
     (3) USB 2 device is connected when entering S3, remove the device,
         and USB mode bit is unset after back to S0.
     (4) Connect a A->C adapter, insert USB 2/3 device at S3, and the
         USB mode bit is set after leaving S3. `lsusb` also identifies
	 device under USB 2/3 bus.
     (5) USB mode bit is always unset if no device is connected.
     (6) type c display is connected, register value not changed.
BRANCH=None

Change-Id: Ic85ee805ffea26bfa3fc5004c3d319cefcf2e68e
Signed-off-by: lschyi <lschyi@google.com>