ap_pwrseq: s5 inactivity timeout handling
current implementaion restarts the s5 inaciver timer after if expires
which delays the S5 to G3 transition.
Problem is seen when the S5 inactive timer expiry is not
read immdediatly after it expires and instaed reading the timer
remaining status which returns zero indicating timer has not started
at all. Hence it starts the inactive timer again.

This is solved by introducing a static global which captures the info
of the timer start in S5 state and avoids restarting the timer again.

BUG=b:235783840
BRANCH=main
TEST=verify S5 to G3 transition time on nivviks

Change-Id: I434a5885ea44b48f6b6b016fc22e56a43161b8ff
Signed-off-by: Deepti Deshatty <deepti.deshatty@intel.corp-partner.google.com>