TCPMV2: Add USB PD Power Button Support
Introduced in USB PD 3.1, Alert messages can now carry information about
power button presses and releases. This CL updates the EC to receive
button press alert messages from the partner and use them to trigger
startup or shutdown events. A short or long press while the system is
off will result in it turning on. A long press while the system is on
will lead to it shutting off.

BUG=b:236022894
TEST=make try_build_boards and make runhosttests, also working
interactivly with a dock that supports sending alerts on button presses.
BRANCH=None

Signed-off-by: Jameson Thies <jthies@google.com>
Change-Id: I425babb537e1fa47962791c0debb7b622fe3c5bf