tree: Files should end with single newline
One of the checks that runs as part of "repo upload" looks for a single
newline at the end of each file. I'm getting warnings about this when I
touch files that do not follow this, even though I didn't add the extra
newlines.

This commit fixes additional files not included in
https://crrev.com/c/3229797 by running the following:

for f in $(find . -name '*.mk');
  do printf '%s\n' "$(cat ${f})" > ${f};
done

for f in $(find . -name '*.S'); do
  printf '%s\n' "$(cat ${f})" > ${f}
done

BRANCH=none
BUG=b:172020503, b:234181908
TEST=./util/compare_build.sh -b all -j 120

Signed-off-by: Tom Hughes <tomhughes@chromium.org>
Change-Id: Ib4e4d0e74bf3b00a0b2c81505fef65e2c1b7401f