zephyr: console: redirect messages to printk when shell is inactive
The log messages printed in the main thread (e.x. the system banner)
cannot be shown in the UART console. Instead, The warning message
"WARNING: A print request was detected on not active shell backend." is
printed. This CL redirects the log message to printk if the shell thread
is inactive.

BRANCH=none
BUG=none
TEST=Check the system banner and other log messages in the main thread
can be printed.

Signed-off-by: Jun Lin <CHLin56@nuvoton.com>
Change-Id: I42365b32d5131a80bd56c27c61e6a9aa360b250d