chip/stm32: Add support for USB forwarding of LPUART
The LPUART is treated as UART9 by the EC codebase, for the purpose of
running the system console.

This CL adds the option of performing USB forwarding on the LPUART, in
the same way as the ordinary numbered UARTS.  In particular, the
LPUART clock divisor formula is different by a factor of 256, which
was already handled in uart.c, but not in usart.c.

BUG=b:192262089
TEST=Observed UART forwarding on HyperDebug
BRANCH=none

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: I1661c7dbbc48ddf556fc14d02e9f438ab2773fcc