common/gpio_commands.c: Consider GPIO reconfiguration
The set() method allows manually setting the level of a named GPIO
pin, provided that the pin is configured for output.  Existing code
considered only the compile-time configuration of the pin.  With this
change, if support has been selected, the method will inspect the
current configuration of the particular GPIO pin.

To be fair, the current state does not usually cause issues, since the
set() method is not called by command_gpio_set() in case
CONFIG_CMD_GPIO_EXTENDED is enabled, which allows reconfiguration of
pins from input to output.

On my board (HyperDebug) I do not want to enable
CONFIG_CMD_GPIO_EXTENDED, but want to keep "gpioset" for only setting
0/1.  At the same time I have implemented a separate custom command
which allows modifying the pin input/output configuration in the ways
needed for the use case.  Without this change, however, I am not able
to use the standard "gpioset" after reconfiguring a pin using my
board-specific command.

BUG=b:192262089
TEST=Use HyperDebug to manipulate the RESET line of AndreiBoard
BRANCH=none

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: Ia66dc01fe3ca2ce4330b09431f6edc85bdeec75d