common: Conditionally support printf %l modifier
The libfp library prints some values with PRIx32 or PRId32 format
specifiers which, in their compilation environment, output "%lx" and
"%ld".

Unfortunately, support for printing longs in EC code was deprecated in
crbug.com/984041 after changing it from it being treated as a
hard-coded 64-bit length.  There was concern that new code using %l
with 32-bit values would be cherry-picked to older branches without
the updated printf. In these cases, the older code would interpret
that %l as 64-bit argument, causing it to over-ingest arguments and
potentially behave in an undefined manner.

Printing 32-bit values with "%l" is safe as long as we can guarantee no
legacy code will attempt to print using "%l" with a 64-bit value. The
logic here is protected by a config flag that is only enabled for FPMCU
and FPMCU doesn't use long running release branches.

A printf test is also added to ensure that only dartmonkey and
bloonchipper boards have long32 enabled.

BRANCH=none
BUG=b:234781655
BUG=b:234143158
TEST=./test/run_device_tests.py -b dartmonkey -t printf
TEST=./test/run_device_tests.py -b bloonchipper -t printf

Signed-off-by: Bobby Casey <bobbycasey@google.com>
Change-Id: If432f507a31cc12a4c5c4bdcd07c6141407bd70d