WIP: test: Move switchcap power good GPIO to emulator
The gpio_mb_power_good and switchcap_pg GPIO pins should be set by the
LN9310 emulator in response to the board enabling and disabling the
switchcap power.

Add an interrupt handler to the gpio_switchcap_on in the LN9310 emulator
that sets the aforementioned GPIOs when switchcap is enabled,
i.e. gpio_switchcap_on is edge triggered.

BRANCH=none
BUG=b:217554681
TEST=zmake test test-drivers

Signed-off-by: Aaron Massey <aaronmassey@google.com>
Change-Id: I6d5736d3ec2fd6189679d31ced67484ec8d6f1dc